Set min map zoom, calculate pace, save run

+++ core/src/main/java/com/karyna/core/domain/LatLng.kt
@@ -1,3 +1,3 @@
 package com.karyna.core.domain
 
-data class LatLng(val lat: Float, val lng: Float)
+data class LatLng(val lat: Double, val lng: Double)
